+++ CHANGELOG.md
@@ -10,7 +10,8 @@ This is a small bug fix release.
 
 ### Changelog
 
-Fix: Lua wrapper for `mympd.api`
+- Fix: Lua wrapper for `mympd.api`
+- Fix: Emmit correct trigger event for INTERNAL_API_TRIGGER_EVENT_EMIT
 
 ***