+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is a small bug fix release.
 
 ### Changelog
 
+- Fix: Import of scripts
 - Fix: Lua wrapper for `mympd.api`
 - Fix: Emmit correct trigger event for INTERNAL_API_TRIGGER_EVENT_EMIT