+++ CHANGELOG.md
@@ -4,7 +4,7 @@ https://github.com/jcorporation/myMPD/
 
 ***
 
-## myMPD v19.0.1 (not yet released)
+## myMPD v19.0.1 (2024-12-15)
 
 This is a small bug fix release.